+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 1. [12684](https://github.com/influxdata/influxdb/pull/12684): Fix mismatch in bucket row and header
 1. [12703](https://github.com/influxdata/influxdb/pull/12703): Allows user to edit note on cell
+1. [12764](https://github.com/influxdata/influxdb/pull/12764): Fix empty state styles in scrapers in org view
 
 ### UI Improvements